MQTT changes, scanner reaction status , HA typo (#7)
* Update README.md

* Update ha-config.md

* Update ha-config.md

* Update ha-config.md

* Update mqtt.cpp

* Update mqtt.cpp

Removed the slash before 'security' in the mqtt topic and changed "honeypot" to "canary"

* Update README.md

* Update README.md

* Further removed honeypot reference

* Reset moved, retain changed

Because the attack was always retained, if you missed a reset message, the attackinprogress will always remain.
Also the ResetAttackState() was initiated after the async connect, but because the connection was not setup yet, the status would not reset. So I moved it to the connection method. I've created a seperate method for setting status, you can now just use a boolean to set or reset.
Now the status will reset when you reboot the ESP.

* Update UI and fixed typo

removed Honeypot reference
removed typo and changed message
changed glance

* Added Attack Info sensor
Added Attack Info Log
Added screenshot

* Update ha-config.md

* Update ha-config.md

* Update ha-config.md

* Added Lovelace image

Co-authored-by: Simon Mingaars <simon.mingaars@pliant.nl>